+++ app/lib/service/entrypoint/frontend.dart
@@ -159,8 +159,9 @@ Future _worker(WorkerEntryMessage message) async {
     Future<void> triggerDependentAnalysis(
         String package, String version, Set<String> affected) async {
       await analyzerClient.triggerAnalysis(package, version, affected);
-      await dartdocClient.triggerDartdoc(package, version,
-          dependentPackages: affected);
+      // TODO: re-enable this after we have added some stop-gaps on the frequency
+      // await dartdocClient.triggerDartdoc(package, version,
+      //    dependentPackages: affected);
     }
 
     final pdb = await PackageDependencyBuilder.loadInitialGraphFromDb(